+++ packages/eslint-plugin/src/__tests__/index.test.ts
@@ -1,0 +1,16 @@
+import * as index from '../index'
+describe('export index', () => {
+  it('export', () => {
+    expect({ ...index }).toEqual({
+      rules: expect.any(Object),
+      configs: {
+        recommended: expect.any(Object),
+      },
+      default: {
+        configs: {
+          recommended: expect.any(Object),
+        },
+      },
+    })
+  })
+})