+++ packages/next-plugin/src/plugin.ts
@@ -1,5 +1,5 @@
-import { existsSync, mkdirSync, writeFileSync } from 'node:fs'
-import { basename, join, resolve } from 'node:path'
+import { existsSync, mkdirSync, readFileSync, writeFileSync } from 'node:fs'
+import { join, relative, resolve } from 'node:path'
 
 import {
   DevupUIWebpackPlugin,
@@ -22,8 +22,14 @@ export function DevupUI(
   config: NextConfig,
   options: DevupUiNextPluginOptions = {},
 ): NextConfig {
-  const isTurbo = process.env.TURBOPACK === '1'
+  const isTurbo =
+    process.env.TURBOPACK === '1' || process.env.TURBOPACK === 'auto'
+  // turbopack is now stable, TURBOPACK is set to auto without any flags
   if (isTurbo) {
+    if (process.env.NODE_ENV === 'production') {
+      throw new Error('Devup UI is not supported in production with turbopack')
+    }
+
     config ??= {}
     config.turbopack ??= {}
     config.turbopack.rules ??= {}
@@ -32,22 +38,31 @@ export function DevupUI(
       distDir = 'df',
       cssDir = resolve(distDir, 'devup-ui'),
       singleCss = false,
+      devupFile = 'devup.json',
     } = options
 
     const sheetFile = join(distDir, 'sheet.json')
     const classMapFile = join(distDir, 'classMap.json')
     const fileMapFile = join(distDir, 'fileMap.json')
     const gitignoreFile = join(distDir, '.gitignore')
-    if (!existsSync(distDir)) mkdirSync(distDir)
-    if (!existsSync(cssDir)) mkdirSync(cssDir)
+    if (!existsSync(distDir))
+      mkdirSync(distDir, {
+        recursive: true,
+      })
+    if (!existsSync(cssDir))
+      mkdirSync(cssDir, {
+        recursive: true,
+      })
     if (!existsSync(gitignoreFile)) writeFileSync(gitignoreFile, '*')
+    // disable turbo parallel
+    process.env.TURBOPACK_DEBUG_JS = '*'
+    process.env.NODE_OPTIONS ??= ''
+    process.env.NODE_OPTIONS += ' --inspect-brk'
+
     const rules: NonNullable<typeof config.turbopack.rules> = {
-      [basename(cssDir)]: [
+      [`./${relative(process.cwd(), cssDir).replaceAll('\\', '/')}/*.css`]: [
         {
           loader: '@devup-ui/webpack-plugin/css-loader',
-          options: {
-            watch: process.env.NODE_ENV === 'development',
-          },
         },
       ],
       '*.{tsx,ts,js,mjs}': [
@@ -61,6 +76,10 @@ export function DevupUI(
             fileMapFile,
             watch: process.env.NODE_ENV === 'development',
             singleCss,
+            // for turbopack, load theme is required on loader
+            theme: existsSync(devupFile)
+              ? JSON.parse(readFileSync(devupFile, 'utf-8'))?.['theme']
+              : undefined,
           },
         },
       ],