+++ packages/reset-css/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @devup-ui/reset-css
 
+## 1.0.13
+
+### Patch Changes
+
+- Updated dependencies [[`8e46b5e`](https://github.com/dev-five-git/devup-ui/commit/8e46b5eb35897ff33b27738f4623ca7bee2588fb)]:
+  - @devup-ui/react@1.0.22
+
+## 1.0.12
+
+### Patch Changes
+
+- Updated dependencies [[`3c65364`](https://github.com/dev-five-git/devup-ui/commit/3c65364125cea6e3582562b99a9b71291fc6f8c2)]:
+  - @devup-ui/react@1.0.21
+
+## 1.0.11
+
+### Patch Changes
+
+- Updated dependencies [[`fee97f8`](https://github.com/dev-five-git/devup-ui/commit/fee97f8a3d79b9a4c62858deb8e1aea8c609e3a2)]:
+  - @devup-ui/react@1.0.20
+
 ## 1.0.10
 
 ### Patch Changes